+++ lib/postgrex/extensions/timetz.ex
@@ -5,17 +5,7 @@ defmodule Postgrex.Extensions.TimeTZ do
 
   @day (:calendar.time_to_seconds({23, 59, 59}) + 1) * 1_000_000
 
-  def init(opts), do: Keyword.fetch!(opts, :date)
-
-  def encode(:postgrex) do
-    quote location: :keep do
-      %Postgrex.Time{} = time ->
-        unquote(__MODULE__).encode_postgrex(time)
-      other ->
-        raise ArgumentError, Postgrex.Utils.encode_msg(other, Postgrex.Time)
-    end
-  end
-  def encode(:elixir) do
+  def encode(_) do
     quote location: :keep do
       %Time{} = time ->
         unquote(__MODULE__).encode_elixir(time)
@@ -24,13 +14,7 @@ defmodule Postgrex.Extensions.TimeTZ do
     end
   end
 
-  def decode(:postgrex) do
-    quote location: :keep do
-      <<12 :: int32, microsecs :: int64, tz :: int32>> ->
-        unquote(__MODULE__).microsecond_to_postgrex(microsecs, tz)
-    end
-  end
-  def decode(:elixir) do
+  def decode(_) do
     quote location: :keep do
       <<12 :: int32, microsecs :: int64, tz :: int32>> ->
         unquote(__MODULE__).microsecond_to_elixir(microsecs, tz)
@@ -38,19 +22,6 @@ defmodule Postgrex.Extensions.TimeTZ do
   end
 
   ## Helpers
-
-  def encode_postgrex(%Postgrex.Time{hour: hour, min: min, sec: sec, usec: usec})
-      when hour in 0..23 and min in 0..59 and sec in 0..59 and usec in 0..999_999 do
-    time = {hour, min, sec}
-    <<12 :: int32, :calendar.time_to_seconds(time) * 1_000_000 + usec :: int64,
-      0 :: int32>>
-  end
-
-  def microsecond_to_postgrex(microsec, tz) do
-    microsec
-    |> adjust_microsecond(tz)
-    |> microsecond_to_postgrex()
-  end
 
   defp adjust_microsecond(microsec, tz) do
     case microsec + tz * 1_000_000 do
@@ -61,13 +32,6 @@ defmodule Postgrex.Extensions.TimeTZ do
       adjusted_microsec ->
         adjusted_microsec - @day
     end
-  end
-
-  defp microsecond_to_postgrex(microsecs) do
-    secs = div(microsecs, 1_000_000)
-    usec = rem(microsecs, 1_000_000)
-    {hour, min, sec} = :calendar.seconds_to_time(secs)
-    %Postgrex.Time{hour: hour, min: min, sec: sec, usec: usec}
   end
 
   def encode_elixir(%Time{hour: hour, minute: min, second: sec, microsecond: {usec, _}})